requirements.txt
crewai
crewai[tools]
together
chromadb
python-dotenv
streamlit
tiktoken
